split jobs definition again in their own files
