+++ pkg/keystone/bootstrap.go
@@ -16,13 +16,17 @@ limitations under the License.
 package keystone
 
 import (
-	keystonev1 "github.com/openstack-k8s-operators/keystone-operator/api/v1beta1"
 	keystonev1beta1 "github.com/openstack-k8s-operators/keystone-operator/api/v1beta1"
 
 	common "github.com/openstack-k8s-operators/lib-common/pkg/common"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+const (
+	// BootstrapCommand -
+	BootstrapCommand = "/usr/local/bin/kolla_set_configs && keystone-manage bootstrap"
 )
 
 // BootstrapJob func
@@ -36,7 +40,7 @@ func BootstrapJob(
 	if instance.Spec.Debug.Bootstrap {
 		args = append(args, DebugCommand)
 	} else {
-		args = append(args, BootStrapCommand)
+		args = append(args, BootstrapCommand)
 	}
 
 	envVars := map[string]common.EnvSetter{}
@@ -102,74 +106,6 @@ func BootstrapJob(
 	}
 	job.Spec.Template.Spec.Containers[0].Env = common.MergeEnvs(job.Spec.Template.Spec.Containers[0].Env, envVars)
 	job.Spec.Template.Spec.Volumes = getVolumes(instance.Name)
-
-	initContainerDetails := APIDetails{
-		ContainerImage: instance.Spec.ContainerImage,
-		DatabaseHost:   instance.Status.DatabaseHostname,
-		DatabaseUser:   instance.Spec.DatabaseUser,
-		DatabaseName:   DatabaseName,
-		OSPSecret:      instance.Spec.Secret,
-		VolumeMounts:   getInitVolumeMounts(),
-	}
-	job.Spec.Template.Spec.InitContainers = initContainer(initContainerDetails)
-
-	return job
-}
-
-// DbSyncJob func
-func DbSyncJob(
-	instance *keystonev1.KeystoneAPI,
-) *batchv1.Job {
-	runAsUser := int64(0)
-
-	labels := map[string]string{
-		"app": "keystone-api",
-	}
-
-	args := []string{"-c"}
-	if instance.Spec.Debug.DBSync {
-		args = append(args, DebugCommand)
-	} else {
-		args = append(args, DBSyncCommand)
-	}
-
-	envVars := map[string]common.EnvSetter{}
-	envVars["KOLLA_CONFIG_FILE"] = common.EnvValue(KollaConfig)
-	envVars["KOLLA_CONFIG_STRATEGY"] = common.EnvValue("COPY_ALWAYS")
-	envVars["KOLLA_BOOTSTRAP"] = common.EnvValue("true")
-
-	job := &batchv1.Job{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      ServiceName + "-db-sync",
-			Namespace: instance.Namespace,
-			Labels:    labels,
-		},
-		Spec: batchv1.JobSpec{
-			Template: corev1.PodTemplateSpec{
-				Spec: corev1.PodSpec{
-					RestartPolicy:      "OnFailure",
-					ServiceAccountName: ServiceAccount,
-					Containers: []corev1.Container{
-						{
-							Name: ServiceName + "-db-sync",
-							Command: []string{
-								"/bin/bash",
-							},
-							Args:  args,
-							Image: instance.Spec.ContainerImage,
-							SecurityContext: &corev1.SecurityContext{
-								RunAsUser: &runAsUser,
-							},
-							Env:          common.MergeEnvs([]corev1.EnvVar{}, envVars),
-							VolumeMounts: getVolumeMounts(),
-						},
-					},
-				},
-			},
-		},
-	}
-
-	job.Spec.Template.Spec.Volumes = getVolumes(ServiceName)
 
 	initContainerDetails := APIDetails{
 		ContainerImage: instance.Spec.ContainerImage,
